Create README
Welcome to my portfolio! I am Zainab Akhtar, a budding software developer currently majoring in Information Technology. This repository serves as a showcase for my projects, where I aim to combine my programming skills with design principles. 

In addition to my passion for coding, I have a strong interest in User Experience (UX) design. I believe that programming is not just about writing code but also about creating user-friendly applications that provide a seamless experience. My goal is to blend technology with thoughtful design, ensuring that the end-user has a positive interaction with the software.

Each project in this portfolio reflects my journey in learning Java, python and web development technologies, and I strive to improve both my coding and design capabilities continuously. I hope this portfolio serves as a demonstration of my dedication and growth in the tech field.